for start

+++ src/components/tuits/index.js
@@ -1,23 +1,62 @@
-import React from "react";
+import React, {useEffect, useState} from "react";
 import './tuits.css';
 import Tuit from "./tuit";
 import * as likesService from "../../services/like-service";
 import * as service from "../../services/tuits-service";
+import * as authService from "../../services/auth-service";
+import {useNavigate} from "react-router-dom";
 
 const Tuits = ({tuits = [], refreshTuits}) => {
-    const likeTuit = (tuit) =>
-        likesService.userTogglesTuitLikes("my", tuit._id)
-            .then(refreshTuits)
-            .catch(e => alert(e));
 
-    const deleteTuit = (tid) =>
-        service.deleteTuit(tid)
-            .then(refreshTuits);
+    const [profile, setProfile] = useState(undefined);
+    const navigate = useNavigate();
 
-    const dislikeTuit = (tuit) =>
-        likesService.userTogglesTuitDislikes("my",tuit._id)
-            .then(refreshTuits)
-            .catch(e=>alert(e));
+    useEffect(async ()=> {
+        try {
+            const user = await authService.profile();
+            if (user) {
+                setProfile(user);
+            }
+        } catch (e) {
+        }
+    }, []);
+
+    const likeTuit = (tuit) =>{
+        if(profile !== undefined){
+            likesService.userTogglesTuitLikes("my", tuit._id)
+                .then(refreshTuits)
+                .catch(e => alert(e));
+        }else{
+            alert("Please log in!");
+            navigate('/login');
+        }
+    }
+
+
+    const deleteTuit = (tid) => {
+        if(profile !== undefined){
+            service.deleteTuit(tid)
+                .then(refreshTuits);
+        }else{
+            alert("Please log in!");
+            navigate('/login');
+        }
+    }
+
+
+    const dislikeTuit = (tuit) => {
+        if(profile!==undefined){
+            likesService.userTogglesTuitDislikes("my",tuit._id)
+                .then(refreshTuits)
+                .catch(e=>alert(e));
+        }else{
+            alert("Please log in!");
+            navigate('/login');
+        }
+
+
+    }
+
 
     return (
         <div>
